build: creating SelectSection component
