+++ package.json
@@ -13,6 +13,7 @@
     "@testing-library/jest-dom": "^5.16.4",
     "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^13.5.0",
+    "axios": "^0.27.2",
     "node-xcs": "^0.1.7",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
