+++ src/pages/Pokemons/index.js
@@ -2,25 +2,13 @@ import { Container } from "@mui/material";
 import { useEffect, useState } from "react";
 import { PokemonsContainer } from "./style";
 import lupa from "./../../img/lupa.png";
-
-import KeyboardArrowDownOutlinedIcon from "@mui/icons-material/KeyboardArrowDownOutlined";
-import KeyboardArrowUpOutlinedIcon from "@mui/icons-material/KeyboardArrowUpOutlined";
 import AllPokemons from "../../components/AllPokemons";
 import ShowPokemonDesck from "../ShowPokemonDesck";
+import SelectSection from "../../components/layout/SelectSection";
 
 export const Pokemons = () => {
-  const [tipo, setTipo] = useState(false);
-  const [ataque, setAtaque] = useState(false);
-  const [defesa, setDefesa] = useState(false);
   const [buscarPokemon, setBuscarPokemon] = useState("");
   const [verPokemon, setVerPokemon] = useState(false);
-
-  const showTipo = () => setTipo(true);
-  const hiderTipo = () => setTipo(false);
-  const showAtaque = () => setAtaque(true);
-  const hiderAtaque = () => setAtaque(false);
-  const showDefesa = () => setDefesa(true);
-  const hiderDefesa = () => setDefesa(false);
 
   const [pokemonsList, setPokemonsList] = useState([]);
   const [loadMore, setLoadMore] = useState(
@@ -69,8 +57,8 @@ export const Pokemons = () => {
     // await console.log(pokemonsList);
   };
 
-  const showNewPokemon = async (pokemon) => {
-    let url = `https://pokeapi.co/api/v2/pokemon/${pokemon}`;
+  const showNewPokemon = async (buscarPokemon) => {
+    let url = `https://pokeapi.co/api/v2/pokemon/${buscarPokemon}`;
     const response = await fetch(url);
     return await response.json();
   };
@@ -79,11 +67,13 @@ export const Pokemons = () => {
     getAllPokemons();
   }, []);
 
-  const openModal = (index) => {
-    //console.log("Deu!");
-    setVerPokemon(index);
-  };
+  //console.log(buscarPokemon);
 
+/*
+Difculdade:
+Para selecionar apenas um pokémon ao clicar para ver no modal;
+Buscar um pokémon na lista da api
+*/
   return (
     <Container>
       <PokemonsContainer>
@@ -100,141 +90,8 @@ export const Pokemons = () => {
             <div className="pokemonsIcon" onClick={showNewPokemon}>
               <img src={lupa} alt="Lupa" />
             </div>
-            {/* TESTE
-            {pokemonsList.map((pokemons, index) => (
-              <>
-                {buscarPokemon === pokemons.name ? (
-                  <ShowPokemonDesck
-                    onClose={() => setVerPokemon(false)}
-                    onClick={() => {
-                      openModal(index.id);
-                    }}
-                    key={index}
-                    id={pokemons.id}
-                    image={pokemons.sprites.other.dream_world.front_default}
-                    name={pokemons.name}
-                    specie={pokemons.species.name}
-                    type={pokemons.types[0].type.name}
-                    experience={pokemons.base_experience}
-                    weight={pokemons.weight}
-                    height={pokemons.height}
-                    abilities={pokemons.abilities[0].ability.name}
-                  ></ShowPokemonDesck>
-                ) : (
-                  console.log(
-                    `O POKÉMON ${buscarPokemon} NÃO FOI ENCONTRADO, \nverifique se o nome está correto!`
-                  )
-                )}
-              </>
-            ))}
-             FIM TESTE */}
           </div>
-          <div className="pokemonsGroupSelect">
-            <div className="pkemonsSelectTipoCheckbox">
-              <div name="tipo" className="pokemonsSelectTipo">
-                <p>Tipo</p>
-                {tipo ? (
-                  <KeyboardArrowUpOutlinedIcon
-                    onClick={hiderTipo}
-                    className="pokemonsIcons"
-                  ></KeyboardArrowUpOutlinedIcon>
-                ) : (
-                  <KeyboardArrowDownOutlinedIcon
-                    onClick={showTipo}
-                    className="pokemonsIcons"
-                  ></KeyboardArrowDownOutlinedIcon>
-                )}
-              </div>
-              {tipo ? (
-                <div className="pokemonsCheckboxTipo">
-                  <input type="checkbox" name="tipo" value="Fogo" />
-                  {" Fogo"}
-                  <br />
-                  <input type="checkbox" name="tipo" value="Planta" />
-                  {" Planta "}
-                  <br />
-                  <input type="checkbox" name="tipo" value="Eletrico" />
-                  {" Eletrico "}
-                  <br />
-                  <input type="checkbox" name="tipo" value="Agua" />
-                  {" Água"}
-                  <br />
-                  <input type="checkbox" name="tipo" value="Normal" />
-                  {" Normal"}
-                </div>
-              ) : null}
-            </div>
-
-            <div className="pkemonsSelectTipoCheckbox">
-              <div name="ataque" className="pokemonsSelectAtaque">
-                <p>Ataque</p>
-                {ataque ? (
-                  <KeyboardArrowUpOutlinedIcon
-                    onClick={hiderAtaque}
-                    className="pokemonsIcons"
-                  ></KeyboardArrowUpOutlinedIcon>
-                ) : (
-                  <KeyboardArrowDownOutlinedIcon
-                    onClick={showAtaque}
-                    className="pokemonsIcons"
-                  ></KeyboardArrowDownOutlinedIcon>
-                )}
-              </div>
-              {ataque ? (
-                <div className="pokemonsCheckboxTipo">
-                  <input type="checkbox" name="ataque" value="Fogo" />
-                  {" Fogo"}
-                  <br />
-                  <input type="checkbox" name="ataque" value="Planta" />
-                  {" Planta "}
-                  <br />
-                  <input type="checkbox" name="ataque" value="Eletrico" />
-                  {" Eletrico "}
-                  <br />
-                  <input type="checkbox" name="ataque" value="Agua" />
-                  {" Água"}
-                  <br />
-                  <input type="checkbox" name="ataque" value="Normal" />
-                  {" Normal"}
-                </div>
-              ) : null}
-            </div>
-
-            <div className="pkemonsSelectTipoCheckbox">
-              <div name="ataque" className="pokemonsSelectDefesa">
-                <p>Defesa</p>
-                {defesa ? (
-                  <KeyboardArrowUpOutlinedIcon
-                    onClick={hiderDefesa}
-                    className="pokemonsIcons"
-                  ></KeyboardArrowUpOutlinedIcon>
-                ) : (
-                  <KeyboardArrowDownOutlinedIcon
-                    onClick={showDefesa}
-                    className="pokemonsIcons"
-                  ></KeyboardArrowDownOutlinedIcon>
-                )}
-              </div>
-              {defesa ? (
-                <div className="pokemonsCheckboxTipo">
-                  <input type="checkbox" name="defesa" value="Fogo" />
-                  {" Fogo"}
-                  <br />
-                  <input type="checkbox" name="defesa" value="Planta" />
-                  {" Planta "}
-                  <br />
-                  <input type="checkbox" name="defesa" value="Eletrico" />
-                  {" Eletrico "}
-                  <br />
-                  <input type="checkbox" name="defesa" value="Agua" />
-                  {" Água"}
-                  <br />
-                  <input type="checkbox" name="defesa" value="Normal" />
-                  {" Normal"}
-                </div>
-              ) : null}
-            </div>
-          </div>
+          <SelectSection />
           <div
             className="pokemonsAllpokemonsBox"
             onClick={() => {
@@ -258,9 +115,7 @@ export const Pokemons = () => {
             {verPokemon ? (
               <ShowPokemonDesck
                 onClose={() => setVerPokemon(false)}
-                onClick={() => {
-                  openModal(index.id);
-                }}
+                onClick={() => setPokemonsList(pokemons.id)}
                 key={index}
                 id={pokemons.id}
                 image={pokemons.sprites.other.dream_world.front_default}
